Made PostgresConfig interoperable with ConnectInfo from the PostgresSimple library
The current structure is a bit of a pain to work with if/when you
actually want PostgresSimple's default behaviour w.r.t default values
and reading from ENV vars and .pgpass files.

This adds a parser for ConnectInfo, and conversion functions between
ConnectInfo and PostgresConfig

Most significantly, I've changed `parseConfig` to use PostgresSimple's
in-built connection string builder, instead of the custom one we had,
which means we get all the fallback logic.